change main page
